+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be logged here
 - ✏️ move global CSS variables from body to :root
 - ✏️ add links to `README.md` for the technologies badges
 - ✏️ move css var from `.dark` and `.light` to `:root` and `:root.light`
+- ✏️ move store selection to `useActiveStore.ts`
 
 ## 0.1.10